generate-blueprint: modularize
