+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/cli/commands.mjs.ejs
