+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/generate-sample/command.mjs.ejs
