+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/generate-sample/generator.mjs.ejs
