+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/generate-sample/index.mjs.ejs
