+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/generate-sample/templates/samples/sample.jdl.ejs
