+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/github-build-matrix/command.mjs.ejs
