+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/github-build-matrix/generator.spec.mjs.ejs
