+++ generators/generate-blueprint/generators/standalone/templates/.blueprint/github-build-matrix/index.mjs.ejs
