+++ generators/generate-blueprint/generators/standalone/templates/.github/workflows/build-cache.yml.ejs
