+++ generators/generate-blueprint/generators/standalone/templates/.github/workflows/generator.yml.ejs
