+++ generators/generate-blueprint/generators/standalone/templates/.github/workflows/samples.yml.ejs
