+++ generators/generate-blueprint/generators/standalone/templates/.gitignore.jhi.blueprint.ejs
