+++ generators/generate-blueprint/generators/standalone/templates/.prettierignore.jhi.blueprint.ejs
