+++ generators/generate-blueprint/generators/standalone/templates/README.md.ejs
