+++ generators/generate-blueprint/generators/standalone/templates/cli/cli-customizations.cjs.ejs
