+++ generators/generate-blueprint/generators/standalone/templates/cli/cli.cjs.ejs
