+++ generators/generate-blueprint/generators/standalone/templates/eslint.config.ts.jhi.blueprint.ejs
