+++ generators/generate-blueprint/generators/standalone/templates/tsconfig.json.ejs
