+++ generators/generate-blueprint/generators/standalone/templates/vitest.config.ts.ejs
