+++ generators/generate-blueprint/generators/standalone/templates/vitest.test-setup.ts.ejs
