+++ generators/generate-blueprint/internal/lookup-namespaces.spec.ts
@@ -55,6 +55,9 @@ describe('lookupGeneratorsNamespaces', () => {
   "entity",
   "export-jdl",
   "generate-blueprint",
+  "generate-blueprint:bootstrap",
+  "generate-blueprint:local",
+  "generate-blueprint:standalone",
   "git",
   "heroku",
   "info",
